Frozen soil infiltration algorithm implemented based on Gray et al, 2001 ('Estimating areal snowmelt infiltration into frozen soils'). This  implementation is a direct translation of the FROZEN module of CRHM (Cold Regions Hydrological Model).
How to use it:
Use the FROZENSOILINFILFLAG to activate the implemented frozen soil infiltration.
Setting FROZENSOILINFILFLAG to 0 preserves the CLASS infiltration scheme which corresponds to unlimited infiltration.
Setting FROZENSOILINFILFLAG to 1 activates the new infiltration algorithm. In this case the opportunity time (time available for snow melting) is calculated by the model.
Setting FROZENSOILINFILFLAG to a value greater than 1 activates the new infiltration algorithm. In this case the opportunity time takes the value of the FROZENSOILINFILFLAG. An approximate value of the opportunity time (unit is in hours) can be taken from one of the CLASS outputs that show snow accumulation time series. For example, if the ground is coverred by snow for five months then the opportunity time will be 5*30*24.


git-svn-id: svn://watflood.uwaterloo.ca/mesh/Branches/MESH_CODE_TEAM/Muluneh/TRUNK2@385 04c54e49-724f-4ac9-b36f-5436cbaa49b4