A fix to the calculation of NMELT, which caused the model to crash if the compiler was checking array bounds AND if the run was started on January 1st.
Here's a detailed explanation: There is a problem with the allocation of t0_ACC in the implementation of the frozen soil infiltration algorithm. In the MESH driver, t0_ACC(NMELT) gets passed into CLASSW. The problem is if the run is started on January 1st… in this case, in the final year of the run, NMELT will be greater than NYEARS (the size of the allocation of t0_ACC at the end of read_initial_inputs.f subroutine).

git-svn-id: svn://watflood.uwaterloo.ca/mesh/TRUNK@511 04c54e49-724f-4ac9-b36f-5436cbaa49b4