fix(io): Fixed missing frames at the end of output files
Updated the simulation end date to be the date of the last frame read from forcing files when the model stops because it has run out of forcing data. This ensures frames are printed for partial days, hours, months and/or years that wouldn't otherwise be triggered because the default check of 'now /= next' is not satisfied.

Signed-off-by: dprincz <17090685+dprincz@users.noreply.github.com>