fix(mesh): Fixed messaging regarding projection information and added NetCDF sync
Fixed a message when a projection is not found in a NetCDF file to print as a warning instead of as an error as this check does not prevent the code from continuing. Added NetCDF sync when writing time-series data or when closing the file to preserve data in the buffer if the simulation exits abnormally (crashes or times out).

Signed-off-by: dprincz <17090685+dprincz@users.noreply.github.com>