feat(svs): Renamed remaining `*.ftn90` extensions to `*.F90` or `*.f90`
Renamed remaining instances of the `*.ftn90` extension to `*.F90` where preprocessing is required or might be required, and to `*.f90` otherwise. These are remnant of previous ECCC proprietary coding wrappers and utilities. All other instances of the extension were removed when `librmn` was updated in r1771 (4f25b35). Updated makefile to deactivate the special rule for the `*.ftn90` extension. Updated makefile to deactivate the special rule for `runsvs_mod` (the dependency is captured by the order of files in makefile.def and similar dependencies aren't otherwise captured in this file).

Preserved the `FTN90PP` and `FTN90PPOPT` variables in makefile and kept them tied to the rules for `*.F90` and `*.f90` files until it can be determined there aren't files remaining that aren't properly identified for preprocessing by `F` or `F90` extension (`-cpp`) or otherwise require special consideration of free-form (`-ffree-form`) or free-form line length (`-ffree-line-length-none`). The other codes should not have CRAY dependencies requiring `-fcray-pointer`. The variables are left as-is to accommodate identifying and correcting such instances than accommodating them by updating the general flags directly.