Tracking the end date of input NetCDF file(s)
This commit provides capability to track the "end" date of NetCDF files.
This includes adding the `end` entry to the `io_file` type, and filling
the information whenever a NetCDF file is read. This feature is needed when
MESH needs to switch to a new NetCDF file, if provided, in a multi-NetCDF
file reading scheme.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>