Checking a multi-NetCDF list of files in textual file format
This commit checks whether the "FORCINGLIST" option is provided in the
"MESH_input_run_options.ini" file, which is a necessary input file for MESH
runs. Only applicable to NetCDF input files, this option overwrites the value
of "fname=path/to/input/forcing.nc" typically provided as one of the arguments
of the "BASINFORCINGFLAG" option in this same `.ini` configuration file.

This option also makes the value of "fname=..." option unnecessary, so users
have the capacity to only provide the list of forcing files in the textual file
that read by the "FORCINGLSIT" option.

The list of forcing files in the target textual file contains paths to the
forcing files that need be read in order. Each NetCDF file path must be
provided in a single line. In other words, the delimiter is the `\n` character.

To comply with MESH's current standard, the file extention to the "FORCINGLIST"
option is assumed to be `.txt`, therefore, no need to provide the extention to
this option.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>