Switching between NetCDF forcing files if enabled
This commit provides the capacity for the driver to switch between
NetCDF input files, when the model reaches to the "end" date of
each NetCDF file.

With each time-step, it keeps track of `ic%now%...` values, and if
it turns out to pass the `forcing_files(i)%series%end` value, it
inquires whether a subsequent NetCDF file is provided in the list
fed to the "MESH_input_run_options.ini"'s `FORCINGLIST` option.

If no more data is available, it sets `ENDDATA` to `true` and
prompts the `while` loop to exit.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>