feat(class): To use user-provided ponding limits regardless of `IWF`
Updated `APREP` to use user-provided ponding limits for bare-ground and snow regardless of the state of `IWF`. Previously, CLASS would compute ponding limits using hard-coded values based on the canopy type with `IWF 0`. This behaviour can be maintained by setting the user-provided ponding limits for bare-ground and snow to zero with `IWF 0`.

Signed-off-by: dprincz <17090685+dprincz@users.noreply.github.com>