fix(io): Restored compatibility for `ROTLATLONG` projection in r2c files
Fixed a missing index when assigning from single-dimension `lat` and `lon` fields for non-dimensional files. Moved the derivation of `lat` and `lon` coordinates for equidistant projections using a lower-left reference to be projection-independent. Added the attributes for the `ROTLATLONG` and `rotated_latitude_longitude` projections when updating the legacy `shd` variable.

Resolves #108.