fix(io): Resolves an error that could occur with seeking past the header in r2c-format files.
When compiled using optimizations with `gcc13` and possibly `gcc12`-`14`, a bad `iostat` is returned or `ierr` seems to not hold a non-zero value, which causes the routine to exit without properly seeking the file. This causes warnings that eventually lead to the file being improperly dimensioned for time-based r2c-format files. Moving the initialization/zeroing of the `ierr` output variable before the call to `rewind` the file seems to resolve this issue in optimized code. This issue would not be encountered when compiled using `debug` symbols and was not present when using older `gcc` versions (e.g., `gcc5`-`11`).

Resolves issue #94.